Move mutation logic to individual runners
